Moved SaveLink to own file. Rewrote sharedFolderPageV2 to includeSharedFolder. Deleted unused files
